// src/app/Services/data-store/model/AgentContacts-api-Model.ts
export  class AgentModel_api{
    contactId : string
    name : string;
    idCode: string;
    producerId: string;
    address1: string;
    address2: string;
    city: string;
    state: string;
    postalCode: string;
    country: string;
    mailAddress1: string;
    mailAddress2: string;
    mailCity: string;
    mailState: string;
    mailPostalCode: string;
    phone: string;
    extension: string;
    fax: string;
    home: string;
    title: string;
    email: string;
    url: string;
    salutation  : string;
    mobile : string;
    phoneOther  : string;
}